+++ public/data/db.json
@@ -374,7 +374,7 @@
     },
     {
       "title": "내게 복종하세요",
-      "cover": "/assets/cover/adultcoverimage2.jpeg",
+      "cover": "/assets/cover/adultcoverimage2.png",
       "author": "소소우",
       "freeEpisode": 3,
       "rating": 4.6,
