+++ public/js/components/index.js
@@ -24,3 +24,5 @@ export { default as BookKeyword } from './webtoonpage/BookKeyword.js';
 export { default as BookIntroduce } from './webtoonpage/BookIntroduce.js';
 export { default as BookReview } from './webtoonpage/BookReview.js';
 export { default as BookSimilar } from './webtoonpage/BookSimilar.js';
+export { default as BookModal } from './webtoonpage/BookModal.js';
+export { default as BookSeriesItem } from './webtoonpage/BookSeriesItem.js';
